Update and rename data_sources/code to data_sources/code/DataWranglingRyan
This is all the code in my IPython notebook. No idea how to share those. Anyways each line break is basically a different function. Those are my breaks in the notebook. It looks a lot nice in my notebook lol. Pretty much every  "function" ends with team_rpi, that is for testing purposes and viewing purposes. You can comment those out.

I will try to add some comments here to make this readable. Currently this is rpi with team name, id, and year for seasons ending in 2015 and 2014.