Create get_sos.py
This is messed up and not finished. You can try it out. cbs.com is missing sos for 2011-12 season and everything before it gets shifted up a year. So this gets everything from 94 to 2010. Also there are a few teams in there with team name ' '. Dont know why cbssports messed up so much.